leanote : refactor for easy testing
