+++ src/common/backend/services/leanote/form.tsx
@@ -3,7 +3,7 @@ import '@ant-design/compatible/assets/index.less';
 import { Input, Checkbox } from 'antd';
 import { FormComponentProps } from '@ant-design/compatible/lib/form';
 import React, { Fragment } from 'react';
-import { LeanoteBackendServiceConfig } from './interface';
+import { LeanoteBackendServiceConfig } from '../../clients/leanote/interface';
 import { FormattedMessage } from 'react-intl';
 import i18n from '@/common/locales';
