+++ src/common/backend/services/leanote/service.ts
@@ -1,34 +1,24 @@
 import { CompleteStatus } from 'common/backend/interface';
 import { DocumentService, CreateDocumentRequest } from '../../index';
-import { IBasicRequestService, IExtendRequestHelper } from '@/service/common/request';
-import { RequestHelper } from '@/service/request/common/request';
+import { IBasicRequestService } from '@/service/common/request';
 import { Container } from 'typedi';
-import showdown from 'showdown';
-import { LeanoteBackendServiceConfig, LeanoteNotebook, LeanoteResponse } from './interface';
+import LeanoteClient from '../../clients/leanote/client';
+import { LeanoteBackendServiceConfig, LeanoteNotebook } from '../../clients/leanote/interface';
 
-const converter = new showdown.Converter();
 /**
  * Document service for self hosted leanote or leanote.com
- *
  */
 export default class LeanoteDocumentService implements DocumentService {
+  private client: LeanoteClient;
   private config: LeanoteBackendServiceConfig;
-  private request: IExtendRequestHelper;
-  private formData: FormData;
-  private imagesCount: number;
 
   /**
    * This extension will need the user and password to connect to leanote and fetch a token
    * You must supply the one of your leanote server.
    */
-  constructor({ leanote_host, email, pwd, token_cached }: LeanoteBackendServiceConfig) {
-    this.config = { leanote_host, email, pwd, token_cached };
-    this.formData = new FormData();
-    this.imagesCount = 0;
-    this.request = new RequestHelper({
-      baseURL: this.config.leanote_host,
-      request: Container.get(IBasicRequestService),
-    });
+  constructor(config: LeanoteBackendServiceConfig) {
+    this.config = config;
+    this.client = new LeanoteClient(config, Container.get(IBasicRequestService));
   }
 
   getId = () => {
@@ -45,16 +35,18 @@ export default class LeanoteDocumentService implements DocumentService {
   };
 
   /**
-   * Use the leanote api to fetch notebook as repository
+   * If not logged, login then fetch notebook as repository
    *
    * @see documentation https://github.com/leanote/leanote/wiki/leanote-api
    */
   getRepositories = async () => {
-    let response = await this._getSyncNotebooks();
+    let response = await this.client.getSyncNotebooks();
+    console.log(response);
     if (response.Msg && response.Msg === 'NOTLOGIN') {
-      await this._login();
-      response = await this._getSyncNotebooks();
+      await this.client.login();
+      response = await this.client.getSyncNotebooks();
     }
+    console.log(response);
     return response.map(function(leanoteNotebook: LeanoteNotebook) {
       return {
         id: leanoteNotebook.NotebookId,
@@ -71,18 +63,7 @@ export default class LeanoteDocumentService implements DocumentService {
    * @see documentation https://github.com/leanote/leanote/wiki/leanote-api
    */
   createDocument = async (info: CreateDocumentRequest): Promise<CompleteStatus> => {
-    this.formData.append('NotebookId', info.repositoryId);
-    this.formData.append('Title', info.title);
-    this.formData.append('Content', converter.makeHtml(info.content));
-    const formData = this.formData;
-    this.formData = new FormData();
-    this.imagesCount = 0;
-    await this.request.postForm<LeanoteResponse>(
-      `/api/note/addNote?token=${this.config.token_cached}`,
-      {
-        data: formData,
-      }
-    );
+    await this.client.createDocument(info);
     return {
       href: `${this.config.leanote_host}`,
     };
@@ -94,31 +75,6 @@ export default class LeanoteDocumentService implements DocumentService {
    * @see documentation https://github.com/leanote/leanote/wiki/leanote-api
    */
   uploadBlob = async (blob: Blob): Promise<string> => {
-    const ext = blob.type.split('/').pop();
-    const filename = `${this.imagesCount}.${ext}`;
-    const localFileId = `${this.imagesCount}`;
-    this.formData.append(`Files[${localFileId}][LocalFileId]`, localFileId);
-    this.formData.append(`Files[${localFileId}][Title]`, filename);
-    this.formData.append(`Files[${localFileId}][Type]`, blob.type);
-    this.formData.append(`Files[${localFileId}][HasBody]`, 'true');
-    this.imagesCount++;
-    this.formData.append(`FileDatas[${localFileId}]`, blob, filename);
-    return `${this.config.leanote_host}/api/file/getImage?fileId=${localFileId}`;
-  };
-
-  _login = async () => {
-    if (!this.config.email || !this.config.pwd || this.config.email === '') {
-      throw new Error('Cannot login');
-    }
-    const data = await this.request.get<LeanoteResponse>(
-      `/api/auth/login?email=${this.config.email}&pwd=${this.config.pwd}`
-    );
-    this.config.token_cached = data.Token;
-  };
-
-  _getSyncNotebooks = () => {
-    return this.request.get<LeanoteNotebook[] | LeanoteResponse>(
-      `/api/notebook/getSyncNotebooks?token=${this.config.token_cached}`
-    );
+    return this.client.uploadBlob(blob);
   };
 }
